refactor(tests): lift retry() into assert_log()
