+++ test/functional/api/server_notifications_spec.lua
@@ -84,9 +84,7 @@ describe('notify', function()
     }}
     nvim('subscribe', 'event1')
     nvim('unsubscribe', 'doesnotexist')
-    retry(nil, 1000, function()
-      assert_log("tried to unsubscribe unknown event 'doesnotexist'", testlog, 10)
-    end)
+    assert_log("tried to unsubscribe unknown event 'doesnotexist'", testlog, 10)
     nvim('unsubscribe', 'event1')
     assert_alive()
   end)
